src/multistack/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
alsa-lib=1.2.13=hb9d3cd8_0
aom=3.9.1=hac33072_0
attr=2.5.1=h166bdaf_1
autograd=1.7.0=pyhd8ed1ab_1
brotli=1.1.0=hb9d3cd8_2
brotli-bin=1.1.0=hb9d3cd8_2
bzip2=1.0.8=h4bc722e_7
c-ares=1.34.4=hb9d3cd8_0
ca-certificates=2025.1.31=hbcca054_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cairo=1.18.2=h3394656_1
contourpy=1.3.1=py311hd18a35c_0
cycler=0.12.1=pyhd8ed1ab_1
cyrus-sasl=2.1.27=h54b06d7_7
dav1d=1.2.1=hd590300_0
dbus=1.13.6=h5008d03_3
double-conversion=3.3.1=h5888daf_0
expat=2.6.4=h5888daf_0
ffmpeg=7.1.0=gpl_h8811464_714
fftw=3.3.10=mpi_mpich_hbcf76dd_10
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_3
fontconfig=2.15.0=h7e30c49_1
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.56.0=py311h2dc5d0c_0
freetype=2.12.1=h267a509_2
fribidi=1.0.10=h36c2ea0_0
gdk-pixbuf=2.42.12=hb9ae30d_0
gettext=0.23.1=h5888daf_0
gettext-tools=0.23.1=h5888daf_0
gmp=6.3.0=hac33072_2
graphite2=1.3.13=h59595ed_1003
gsl=2.7=he838d99_0
h5py=3.10.0=mpi_mpich_py311hef8708c_0
h5utils=1.13.2=mpi_mpich_h0b2737e_1002
harfbuzz=10.3.0=h76408a6_0
harminv=1.4.2=hc567f7b_0
hdf5=1.14.2=mpi_mpich_ha2c2bf8_0
icu=75.1=he02047a_0
jack=1.9.22=h7c63dc7_2
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.7=py311hd18a35c_0
krb5=1.21.3=h659f571_0
lame=3.100=h166bdaf_1003
lcms2=2.17=h717163a_0
ld_impl_linux-64=2.43=h712a8e2_4
lerc=4.0.0=h27087fc_0
level-zero=1.20.6=h84d6215_0
libabseil=20240722.0=cxx17_hbbce691_4
libaec=1.1.3=h59595ed_0
libasprintf=0.23.1=h8e693c7_0
libasprintf-devel=0.23.1=h8e693c7_0
libass=0.17.3=hba53ac1_1
libblas=3.9.0=31_h59b9bed_openblas
libbrotlicommon=1.1.0=hb9d3cd8_2
libbrotlidec=1.1.0=hb9d3cd8_2
libbrotlienc=1.1.0=hb9d3cd8_2
libcap=2.71=h39aace5_0
libcblas=3.9.0=31_he106b2a_openblas
libclang-cpp19.1=19.1.7=default_hb5137d0_1
libclang13=19.1.7=default_h9c6a7e4_1
libctl=4.5.1=h7f98852_0
libcups=2.3.3=h4637d8d_4
libcurl=8.12.1=h332b0f4_0
libdb=6.2.32=h9c3ff4c_0
libdeflate=1.23=h4ddbbb0_0
libdrm=2.4.124=hb9d3cd8_0
libedit=3.1.20250104=pl5321h7949ede_0
libegl=1.7.0=ha4b6fd6_2
libev=4.33=hd590300_2
libexpat=2.6.4=h5888daf_0
libfabric=2.0.0=ha770c72_1
libfabric1=2.0.0=h14e6f36_1
libffi=3.4.6=h2dba641_0
libflac=1.4.3=h59595ed_0
libgcc=14.2.0=h767d61c_2
libgcc-ng=14.2.0=h69a702a_2
libgcrypt-lib=1.11.0=hb9d3cd8_2
libgdsii=0.21=h4bd325d_4
libgettextpo=0.23.1=h5888daf_0
libgettextpo-devel=0.23.1=h5888daf_0
libgfortran=14.2.0=h69a702a_2
libgfortran-ng=14.2.0=h69a702a_2
libgfortran5=14.2.0=hf1ad2bd_2
libgl=1.7.0=ha4b6fd6_2
libglib=2.82.2=h2ff4ddf_1
libglvnd=1.7.0=ha4b6fd6_2
libglx=1.7.0=ha4b6fd6_2
libgomp=14.2.0=h767d61c_2
libgpg-error=1.51=hbd13f7d_1
libhwloc=2.11.2=default_h0d58e46_1001
libiconv=1.18=h4ce23a2_1
libjpeg-turbo=3.0.0=hd590300_1
liblapack=3.9.0=31_h7ac8fdf_openblas
libllvm19=19.1.7=ha7bfdaf_1
liblzma=5.6.4=hb9d3cd8_0
libnghttp2=1.64.0=h161d5f1_0
libnl=3.11.0=hb9d3cd8_0
libnsl=2.0.1=hd590300_0
libntlm=1.8=hb9d3cd8_0
libogg=1.3.5=h4ab18f5_0
libopenblas=0.3.29=pthreads_h94d23a6_0
libopengl=1.7.0=ha4b6fd6_2
libopenvino=2025.0.0=hdc3f47d_1
libopenvino-auto-batch-plugin=2025.0.0=h4d9b6c2_1
libopenvino-auto-plugin=2025.0.0=h4d9b6c2_1
libopenvino-hetero-plugin=2025.0.0=h981d57b_1
libopenvino-intel-cpu-plugin=2025.0.0=hdc3f47d_1
libopenvino-intel-gpu-plugin=2025.0.0=hdc3f47d_1
libopenvino-intel-npu-plugin=2025.0.0=hdc3f47d_1
libopenvino-ir-frontend=2025.0.0=h981d57b_1
libopenvino-onnx-frontend=2025.0.0=h6363af5_1
libopenvino-paddle-frontend=2025.0.0=h6363af5_1
libopenvino-pytorch-frontend=2025.0.0=h5888daf_1
libopenvino-tensorflow-frontend=2025.0.0=h630ec5c_1
libopenvino-tensorflow-lite-frontend=2025.0.0=h5888daf_1
libopus=1.3.1=h7f98852_1
libpciaccess=0.18=hd590300_0
libpng=1.6.47=h943b412_0
libpq=17.4=h27ae623_0
libprotobuf=5.28.3=h6128344_1
librsvg=2.58.4=h49af25d_2
libsndfile=1.2.2=hc60ed4a_1
libsqlite=3.49.1=hee588c1_1
libssh2=1.11.1=hf672d98_0
libstdcxx=14.2.0=h8f9b012_2
libstdcxx-ng=14.2.0=h4852527_2
libsystemd0=257.3=h3dc2cb9_0
libtiff=4.7.0=hd9ff511_3
libudev1=257.3=h9a4d06a_0
libunwind=1.6.2=h9c3ff4c_0
liburing=2.9=h84d6215_0
libusb=1.0.27=h520f47e_100
libuuid=2.38.1=h0b41bf4_0
libva=2.22.0=h8a09558_1
libvorbis=1.3.7=h9c3ff4c_0
libvpx=1.14.1=hac33072_0
libwebp-base=1.5.0=h851e524_0
libxcb=1.17.0=h8a09558_0
libxcrypt=4.4.36=hd590300_1
libxkbcommon=1.8.0=hc4a0caf_0
libxml2=2.13.6=h8d12d68_0
libxslt=1.1.39=h76b75d6_0
libzlib=1.3.1=hb9d3cd8_2
lz4-c=1.10.0=h5888daf_1
matplotlib=3.10.0=py311h38be061_0
matplotlib-base=3.10.0=py311h2b939e6_0
mpb=1.11.1=mpi_mpich_h5eb1b22_4
mpg123=1.32.9=hc50e24c_0
mpi=1.0.1=mpich
mpi4py=4.0.3=py311h0799f86_0
mpich=4.3.0=h1a8bee6_100
munkres=1.1.4=pyh9f0ad1d_0
mysql-common=9.0.1=h266115a_4
mysql-libs=9.0.1=he0572af_4
ncurses=6.5=h2d0b736_3
nlopt=2.10.0=py311h019bca9_0
numpy=1.26.4=py311h64a7726_0
ocl-icd=2.3.2=hb9d3cd8_2
opencl-headers=2024.10.24=h5888daf_0
openh264=2.6.0=hc22cd8d_0
openjpeg=2.5.3=h5fbd93e_0
openldap=2.6.9=he970967_0
openssl=3.4.1=h7b32b05_0
packaging=24.2=pyhd8ed1ab_2
pango=1.56.1=h861ebed_0
pcre2=10.44=hba22ea6_2
pillow=11.1.0=py311h1322bbf_0
pip=25.0.1=pyh8b19718_0
pixman=0.44.2=h29eaf8c_0
pthread-stubs=0.4=hb9d3cd8_1002
pugixml=1.15=h3f63f65_0
pulseaudio-client=17.0=hb77b528_0
pymeep=1.29.0=mpi_mpich_py311h1234567_3
pymeep-extras=1.29.0=mpi_mpich_py311h23d43cc_3
pyparsing=3.2.1=pyhd8ed1ab_0
pyside6=6.8.2=py311h9053184_0
python=3.11.11=h9e4cc4f_1_cpython
python-dateutil=2.9.0.post0=pyhff2d567_1
python_abi=3.11=5_cp311
qhull=2020.2=h434a139_5
qt6-main=6.8.2=h588cce1_0
rdma-core=56.0=h5888daf_0
readline=8.2=h8c095d6_2
scipy=1.15.2=py311h8f841c2_0
sdl2=2.32.50=h9b8e6db_0
sdl3=3.2.4=h3083f51_0
setuptools=75.8.0=pyhff2d567_0
six=1.17.0=pyhd8ed1ab_0
snappy=1.2.1=h8bd8927_1
svt-av1=3.0.0=h5888daf_0
tbb=2022.0.0=hceb3a55_0
tk=8.6.13=noxft_h4845f30_101
tornado=6.4.2=py311h9ecbd09_0
tzdata=2025a=h78e105d_0
ucx=1.18.0=hfd9a62f_1
unicodedata2=16.0.0=py311h9ecbd09_0
wayland=1.23.1=h3e06ad9_0
wayland-protocols=1.41=hd8ed1ab_0
wheel=0.45.1=pyhd8ed1ab_1
x264=1!164.3095=h166bdaf_2
x265=3.5=h924138e_3
xcb-util=0.4.1=hb711507_2
xcb-util-cursor=0.1.5=hb9d3cd8_0
xcb-util-image=0.4.0=hb711507_2
xcb-util-keysyms=0.4.1=hb711507_0
xcb-util-renderutil=0.3.10=hb711507_0
xcb-util-wm=0.4.2=hb711507_0
xkeyboard-config=2.43=hb9d3cd8_0
xorg-libice=1.1.2=hb9d3cd8_0
xorg-libsm=1.2.5=he73a12e_0
xorg-libx11=1.8.11=h4f16b4b_0
xorg-libxau=1.0.12=hb9d3cd8_0
xorg-libxcomposite=0.4.6=hb9d3cd8_2
xorg-libxcursor=1.2.3=hb9d3cd8_0
xorg-libxdamage=1.1.6=hb9d3cd8_0
xorg-libxdmcp=1.1.5=hb9d3cd8_0
xorg-libxext=1.3.6=hb9d3cd8_0
xorg-libxfixes=6.0.1=hb9d3cd8_0
xorg-libxi=1.8.2=hb9d3cd8_0
xorg-libxrandr=1.5.4=hb9d3cd8_0
xorg-libxrender=0.9.12=hb9d3cd8_0
xorg-libxscrnsaver=1.2.4=hb9d3cd8_0
xorg-libxtst=1.2.5=hb9d3cd8_3
xorg-libxxf86vm=1.1.6=hb9d3cd8_0
zlib=1.3.1=hb9d3cd8_2
zstd=1.5.7=hb8e6e7a_1
